fixing something

+++ Chis_Adelin/Week-02/Domain/Domain/CreateMenuItemOp/CreateMenuItemCmd.cs
@@ -11,8 +11,8 @@ namespace Domain.Domain.CreateMenuItemOp
         public string Name { get; }
         public string Ingredients { get; }
         public string Allergens { get; }
-        public double Price { get; }
-        public CreateMenuItemCmd(Menu menu, string name, string ingredients, string allergens, double price)
+        public uint Price { get; }
+        public CreateMenuItemCmd(Menu menu, string name, string ingredients, string allergens, uint price)
         {
             Menu = menu;
             Name = name;
