+++ Chis_Adelin/Week-02/Domain/Models/Order.cs
@@ -8,8 +8,8 @@ namespace Domain.Models
     {
         public DateTime DateTime { get; }
         public List<Menu> Menus{ get; }
-        public double Price { get; }
-        public Order(DateTime dateTime, List<Menu> menus, double price)
+        public uint Price { get; }
+        public Order(DateTime dateTime, List<Menu> menus, uint price)
         {
             DateTime = dateTime;
             Menus = menus;
